requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.4
django-environ==0.4.5
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.7.7
python-baseconv==1.2.2
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
ua-parser==0.10.0
user-agents==2.2.0
whitenoise==5.2.0
